TASK: TASKS.md:2110-2132 - 64-bit clean ABI type definitions and syscall fixes
- Renamed abi-amd64.md to abi-x86_64.md with full utsname table
- Updated ino_t to uint64_t in kernel and userspace types.h
- Updated utsname struct buffers from 65 to 256 bytes
- Fixed sys_lseek declaration to use split 64-bit offset signature
- Updated syscall trace format for lseek (4 args)
- Clarified TASKS.md: native syscalls ARE 64-bit (no _64 suffix pattern)

Build verified: kernel and libc compile without errors.

+++ sys/include/sys/types.h
@@ -14,7 +14,7 @@ typedef long fpos_t;
 
 typedef int32_t mode_t;
 typedef uint32_t dev_t;
-typedef uint32_t ino_t;
+typedef uint64_t ino_t;
 typedef uint32_t nlink_t;
 typedef uint32_t blksize_t;
 typedef int64_t blkcnt_t;
